requirements.txt
drf-yasg==1.21.5
djangorestframework==3.14.0
Django==4.2.1